最優先　リファクタリング #24 refa
